pyproject.toml
[project]
name = "im2deep"
description = "Framework for prediction of collisional cross-section of peptides."
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["proteomics", "peptide", "timsTOF", "CCS"]
authors = [
    { name = "Robbe Devreese", email = "robbe.devreese@ugent.be" },
    { name = "Robbin Bouwmeester", email = "robbin.bouwmeester@ugent.be" },
    { name = "Ralf Gabriels", email = "ralf@gabriels.dev" },
]
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dynamic = ["version"]
requires-python = ">=3.8"
dependencies = [
    "click",
    "deeplc",
    "deeplcretrainer",
    "psm_utils",
    "tensorflow>=2.2,<2.13.0"
]

[project.optional-dependencies]
dev = ["black", "isort>5", "pytest", "pytest-cov"]
docs = [
    "sphinx",
    "numpydoc>=1,<2",
    "recommonmark",
    "sphinx-mdinclude",
    "toml",
    "semver>=2",
    "sphinx_rtd_theme",
    "sphinx-autobuild",
]

[project.urls]
GitHub = "https://github.com/compomics/psm_utils"
ReadTheDocs = "https://psm-utils.readthedocs.io"
PyPi = "https://pypi.org/project/psm-utils/"
CompOmics = "https://www.compomics.com"

[project.scripts]
im2deep = "im2deep.__main__:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"

[tool.black]
line-length = 99
target-version = ['py38']

[tool.ruff]
line-length = 99
target-version = "py38"
